(#698) fix errors from merge and tests

+++ src/hyperion/parameters/components.py
@@ -180,8 +180,8 @@ class TemporaryIspybExtras(BaseModel):
         arbitrary_types_allowed = True
         extra = Extra.forbid
 
-    microns_per_pixel_x: int
-    microns_per_pixel_y: int
+    microns_per_pixel_x: float
+    microns_per_pixel_y: float
     position: list[float] | NDArray
     beam_size_x: float
     beam_size_y: float
